+++ CHANGELOG.md
@@ -7,6 +7,8 @@ We follow [Semantic Versions](https://semver.org/).
 - Improve `django.createsuperuser` command with verbose name settings
 - Fix k8s.set-context when user is not logged in for first time
 - Add pip invocations
+- Add mypy invocations
+- Add pytest invocations
 
 ## 0.7.1